# perf(cubesql): Avoid allocations in MetaContext methods