--- a/rust/cubesql/cubesql/src/compile/rewrite/converter.rs
+++ b/rust/cubesql/cubesql/src/compile/rewrite/converter.rs
@@ -1551,14 +1551,13 @@ impl LanguageToLogicalPlanConverter {
                                 match_data_node!(node_by_id, measure_params[0], MeasureName);
                             let expr = self.to_expr(measure_params[1])?;
                             query_measures.push(measure.to_string());
-                            let data_type = self
-                                .cube_context
-                                .meta
-                                .find_df_data_type(measure.to_string())
-                                .ok_or(CubeError::internal(format!(
-                                    "Can't find measure '{}'",
-                                    measure
-                                )))?;
+                            let data_type =
+                                self.cube_context.meta.find_df_data_type(&measure).ok_or(
+                                    CubeError::internal(format!(
+                                        "Can't find measure '{}'",
+                                        measure
+                                    )),
+                                )?;
                             fields.push((
                                 DFField::new(
                                     expr_relation(&expr),
@@ -1608,14 +1607,13 @@ impl LanguageToLogicalPlanConverter {
                         LogicalPlanLanguage::Dimension(params) => {
                             let dimension = match_data_node!(node_by_id, params[0], DimensionName);
                             let expr = self.to_expr(params[1])?;
-                            let data_type = self
-                                .cube_context
-                                .meta
-                                .find_df_data_type(dimension.to_string())
-                                .ok_or(CubeError::internal(format!(
-                                    "Can't find dimension '{}'",
-                                    dimension
-                                )))?;
+                            let data_type =
+                                self.cube_context.meta.find_df_data_type(&dimension).ok_or(
+                                    CubeError::internal(format!(
+                                        "Can't find dimension '{}'",
+                                        dimension
+                                    )),
+                                )?;
                             query_dimensions.push(dimension.to_string());
                             fields.push((
                                 DFField::new(