--- a/rust/cubesql/cubesql/src/compile/rewrite/converter.rs
+++ b/rust/cubesql/cubesql/src/compile/rewrite/converter.rs
@@ -1705,7 +1705,7 @@ impl LanguageToLogicalPlanConverter {
                                         if self
                                             .cube_context
                                             .meta
-                                            .is_synthetic_field(column.member_name().to_string())
+                                            .is_synthetic_field(column.member_name())
                                         {
                                             fields.push((
                                                 DFField::new(