**Check List**
- [x] Tests have been run in packages where changes made if available
- [x] Linter has been run for changed code
- [x] Tests for the changes have been added if not covered yet
- [x] Docs have been added / updated if required

**Description of Changes Made (if issue reference is not provided)**

Selected results from my laptop:

```
power_bi_wrap           time:   [29.544 ms 29.578 ms 29.616 ms]
                        change: [-21.848% -21.635% -21.387%] (p = 0.00 < 0.05)

long_in_expr            time:   [20.256 ms 20.314 ms 20.348 ms]
                        change: [-4.7778% -4.3730% -4.0255%] (p = 0.00 < 0.05)

quicksight_1            time:   [9.3391 ms 9.3494 ms 9.3598 ms]
                        change: [-4.8924% -4.6718% -4.4695%] (p = 0.00 < 0.05)

large_model_300_select_wildcard
                        time:   [11.278 ms 11.365 ms 11.490 ms]
                        change: [-48.647% -48.052% -47.424%] (p = 0.00 < 0.05)
```